+++ od-win32/picasso96_win.h
@@ -624,8 +624,8 @@ struct picasso96_state_struct
     uae_u16             VirtualHeight; /* Total screen height */
     uae_u8              GC_Depth;    /* From SetGC() */
     uae_u8              GC_Flags;    /* From SetGC() */
-    long                XOffset;     /* From SetPanning() */
-    long                YOffset;     /* From SetPanning() */
+    int                 XOffset;     /* From SetPanning() */
+    int                 YOffset;     /* From SetPanning() */
     uae_u8              SwitchState; /* From SetSwitch() - 0 is Amiga, 1 isPicasso */
     uae_u8              BytesPerPixel;
     uae_u8              CardFound;
@@ -637,7 +637,7 @@ struct picasso96_state_struct
     // support NO direct access all the time to gfx Card
     // every time windows can remove your surface from card so the mainrender place
     // must be in memory
-    long		XYOffset;
+    int         XYOffset;
     bool        dualclut, advDragging;
     int         HLineDBL, VLineDBL;
 };
